+++ src/main/java/umc/unimade/domain/products/controller/ProductsController.java
@@ -13,8 +13,8 @@ import umc.unimade.domain.products.service.ProductsCommandService;
 import umc.unimade.domain.products.service.ProductsQueryService;
 import umc.unimade.global.common.ApiResponse;
 import umc.unimade.global.common.ErrorCode;
-import umc.unimade.global.common.exception.ProductsExceptionHandler;
-import umc.unimade.global.common.exception.UserExceptionHandler;
+import umc.unimade.domain.products.exception.ProductsExceptionHandler;
+import umc.unimade.domain.accounts.exception.UserExceptionHandler;
 
 
 @RestController
