+++ src/main/java/umc/unimade/domain/products/exception/ProductsExceptionHandler.java
@@ -1,0 +1,11 @@
+package umc.unimade.domain.products.exception;
+
+import umc.unimade.global.common.BaseErrorCode;
+import umc.unimade.global.common.exception.CustomException;
+
+public class ProductsExceptionHandler extends CustomException {
+
+    public ProductsExceptionHandler (BaseErrorCode code ){
+        super(code);
+    }
+}
