+++ src/main/java/umc/unimade/domain/products/service/ProductRegisterQueryService.java
@@ -10,7 +10,7 @@ import umc.unimade.domain.products.dto.ProductRegistersResponse;
 import umc.unimade.domain.products.entity.ProductRegister;
 import umc.unimade.domain.products.repository.ProductRegisterRepository;
 import umc.unimade.global.common.ErrorCode;
-import umc.unimade.global.common.exception.ProductsExceptionHandler;
+import umc.unimade.domain.products.exception.ProductsExceptionHandler;
 import umc.unimade.global.registerStatus.RegisterStatus;
 
 @Service
