+++ src/main/java/umc/unimade/domain/review/service/ReviewQueryService.java
@@ -7,7 +7,7 @@ import umc.unimade.domain.review.entity.Review;
 import umc.unimade.domain.review.repository.ReviewRepository;
 import umc.unimade.domain.review.dto.ReviewResponse;
 import umc.unimade.global.common.ErrorCode;
-import umc.unimade.global.common.exception.ReviewExceptionHandler;
+import umc.unimade.domain.review.exception.ReviewExceptionHandler;
 
 @Service
 @Transactional(readOnly = true)
