--- src/main/java/umc/unimade/global/common/exception/ProductsExceptionHandler.java
@@ -1,10 +1,0 @@
-package umc.unimade.global.common.exception;
-
-import umc.unimade.global.common.BaseErrorCode;
-
-public class ProductsExceptionHandler extends CustomException{
-
-    public ProductsExceptionHandler (BaseErrorCode code ){
-        super(code);
-    }
-}
