--- src/main/java/umc/unimade/global/common/exception/QnAExceptionHandler.java
@@ -1,10 +1,0 @@
-package umc.unimade.global.common.exception;
-
-import umc.unimade.global.common.BaseErrorCode;
-
-public class QnAExceptionHandler extends CustomException{
-
-    public QnAExceptionHandler (BaseErrorCode code ){
-        super(code);
-    }
-}
