--- src/main/java/umc/unimade/global/common/exception/ReviewExceptionHandler.java
@@ -1,10 +1,0 @@
-package umc.unimade.global.common.exception;
-
-import umc.unimade.global.common.BaseErrorCode;
-
-public class ReviewExceptionHandler extends CustomException{
-
-    public ReviewExceptionHandler (BaseErrorCode code ){
-        super(code);
-    }
-}
